requirements.txt
# ============================================================
# YouTube Shorts OS — Python Dependencies
# Python 3.11+
# ============================================================

# --- AI / LLM ---
anthropic>=0.40.0              # Claude API for idea gen, scripting, metadata

# --- YouTube APIs ---
google-api-python-client>=2.110.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0

# --- TTS ---
elevenlabs>=1.0.0              # Voiceover generation

# --- Video / Audio ---
moviepy>=1.0.3                 # Video assembly (wraps FFmpeg)
ffmpeg-python>=0.2.0           # Direct FFmpeg bindings for advanced ops
Pillow>=10.0.0                 # Image manipulation for overlays
numpy>=1.26.0                  # Array ops for video frames

# --- HTTP / Scraping ---
httpx>=0.27.0                  # Async HTTP (research engine)
beautifulsoup4>=4.12.0         # HTML parsing
lxml>=5.0.0                    # Fast XML/HTML parser

# --- Database ---
# SQLite is stdlib — no extra package needed
# For future Postgres migration:
# psycopg2-binary>=2.9.0

# --- Config / Settings ---
pydantic>=2.6.0                # Settings validation
pydantic-settings>=2.2.0       # .env loading
python-dotenv>=1.0.0           # .env file support

# --- Scheduling ---
APScheduler>=3.10.0            # Cron-style job scheduler

# --- Dashboard ---
streamlit>=1.35.0              # Control panel UI
plotly>=5.20.0                 # Charts for dashboard
pandas>=2.2.0                  # Data manipulation for analytics

# --- Utilities ---
rich>=13.7.0                   # Terminal output formatting
click>=8.1.0                   # CLI interface for scripts
tenacity>=8.3.0                # Retry logic for API calls
python-slugify>=8.0.0          # URL-safe slugs for filenames
loguru>=0.7.0                  # Structured logging

# --- Dev / Testing ---
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=5.0.0
ruff>=0.4.0                    # Linting + formatting
